% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DashColorscales}
\alias{DashColorscales}
\title{DashColorscales dash-colorscales component}
\usage{
DashColorscales(id = NULL, colorscale = NULL, nSwatches = NULL,
  fixSwatches = NULL, setProps = NULL)
}
\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{colorscale}{Optional: Initial colorscale to display. Default is Viridis.}

\item{nSwatches}{Optional: Initial number of colors in scale to display.}

\item{fixSwatches}{Optional: Set to `True` to fix the number of colors in the scale.}

\item{setProps}{Dash-assigned callback that should be called whenever any of the properties change}
}
\description{
DashColorscales is a Dash wrapper for `react-colorscales`. It takes an array of colors, `colorscale`, and displays a UI for modifying it or choosing a new scale.
}
